requirements.txt
requests
beautifulsoup4
selenium
google-generativeai
Pillow